+++ functions/index.js
@@ -3,7 +3,7 @@
 const express = require("express");
 const { onRequest } = require("firebase-functions/v2/https");
 const { defineSecret } = require("firebase-functions/params");
-const { TrialServiceError, startTrial, verifyTrial } = require("./trialService");
+const { CODES, TrialServiceError, responseBody, startTrial, verifyTrial } = require("./trialService");
 
 const JWT_SECRET = defineSecret("JWT_SECRET");
 
@@ -18,17 +18,25 @@ function getRequestIp(req) {
 
 function sendError(res, error) {
   if (error instanceof TrialServiceError) {
-    return res.status(error.statusCode).json({
-      error: error.message,
-      code: error.code,
-    });
+    return res.status(error.httpStatus).json(
+      responseBody({
+        message: error.message,
+        token: "",
+        statusCode: error.statusCode,
+        error: error.error,
+      })
+    );
   }
 
   console.error("Unhandled error:", error);
-  return res.status(500).json({
-    error: "Internal server error",
-    code: "INTERNAL_ERROR",
-  });
+  return res.status(500).json(
+    responseBody({
+      message: "Internal server error",
+      token: "",
+      statusCode: CODES.INTERNAL_ERROR,
+      error: "INTERNAL_ERROR",
+    })
+  );
 }
 
 function createBaseApp() {
@@ -53,10 +61,14 @@ startTrialApp.post("/", async (req, res) => {
 });
 
 startTrialApp.all("*", (req, res) => {
-  res.status(405).json({
-    error: "Method not allowed",
-    code: "METHOD_NOT_ALLOWED",
-  });
+  res.status(405).json(
+    responseBody({
+      message: "Method not allowed",
+      token: "",
+      statusCode: 4050,
+      error: "METHOD_NOT_ALLOWED",
+    })
+  );
 });
 
 const verifyTrialApp = createBaseApp();
@@ -73,10 +85,14 @@ verifyTrialApp.post("/", async (req, res) => {
 });
 
 verifyTrialApp.all("*", (req, res) => {
-  res.status(405).json({
-    error: "Method not allowed",
-    code: "METHOD_NOT_ALLOWED",
-  });
+  res.status(405).json(
+    responseBody({
+      message: "Method not allowed",
+      token: "",
+      statusCode: 4050,
+      error: "METHOD_NOT_ALLOWED",
+    })
+  );
 });
 
 exports.startTrial = onRequest(